Move PommeProjectionOperator to a pomme module
