+++ docs/api/mapmaking/pomme.md
@@ -1,0 +1,3 @@
+# Pomme
+
+::: furax.mapmaking.pomme
